# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Water Filter Maintenance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /filters:
    get:
      operationId: listFilters
      tags: [filters]
      summary: List all filters
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Filter"
    post:
      operationId: createFilter
      tags: [filters]
      summary: Create a filter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filter"
...
  /filters/upcoming:
    get:
      operationId: getUpcomingReplacements
      tags: [filters]
      summary: Get cartridges due for replacement soon
      parameters:
        - name: withinDays
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UpcomingReplacement"
  /filters/{filterId}:
    get:
      operationId: getFilter
      tags: [filters]
      summary: Get a filter
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filter"
        "404":
          description: Not found
    patch:
      operationId: updateFilter
      tags: [filters]
      summary: Update a filter
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filter"
    delete:
      operationId: deleteFilter
      tags: [filters]
      summary: Delete a filter
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
...
  /filters/{filterId}/cartridges/{cartridgeId}/replace:
    post:
      operationId: replaceCartridge
      tags: [cartridges]
      summary: Mark a cartridge as replaced
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
        - name: cartridgeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cartridge"
  /filters/{filterId}/replace-all:
    post:
      operationId: replaceAllCartridges
      tags: [cartridges]
      summary: Replace all cartridges in a filter at once
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cartridge"
...
components:
  schemas:
...
    Filter:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        location:
          type: ["string", "null"]
        templateType:
          type: string
          enum: [single, three_stage, five_stage, seven_stage, custom]
        installationDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [healthy, warning, overdue]
        cartridges:
          type: array
          items:
            $ref: "#/components/schemas/Cartridge"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, templateType, status, createdAt, updatedAt]
    FilterInput:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        name:
          type: string
          minLength: 1
        location:
          type: ["string", "null"]
        templateType:
          type: string
          enum: [single, three_stage, five_stage, seven_stage, custom]
        installationDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [name, templateType]
    FilterUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: ["string", "null"]
        installationDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    UpcomingReplacement:
      type: object
      properties:
        filterId:
          type: integer
        filterName:
          type: string
        cartridgeId:
          type: integer
        cartridgeName:
          type: string
        stageNumber:
          type: integer
        nextReplacementDate:
          type: string
        daysRemaining:
          type: integer
        status:
          type: string
          enum: [healthy, warning, overdue]
      required: [filterId, filterName, cartridgeId, cartridgeName, stageNumber, nextReplacementDate, daysRemaining, status]
    Cartridge:
      type: object
      properties:
        id:
          type: integer
        filterId:
          type: integer
        name:
          type: string
        stageNumber:
          type: integer
        lastReplacedDate:
          type: ["string", "null"]
        intervalDays:
          type: integer
        nextReplacementDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [healthy, warning, overdue]
        daysRemaining:
          type: ["integer", "null"]
      required: [id, filterId, name, stageNumber, intervalDays, status]
...
    ReplaceInput:
      type: object
      properties:
        replacedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...