# lib/api-zod/src/generated/types/filter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Filter Maintenance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Cartridge } from './cartridge';
import type { FilterStatus } from './filterStatus';
import type { FilterTemplateType } from './filterTemplateType';

export interface Filter {
  id: number;
  /** @nullable */
  userId?: number | null;
  name: string;
  /** @nullable */
  location?: string | null;
  templateType: FilterTemplateType;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  notes?: string | null;
  status: FilterStatus;
  cartridges?: Cartridge[];
  createdAt: string;
  updatedAt: string;
}