# lib/api-zod/src/generated/types/upcomingReplacement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Filter Maintenance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UpcomingReplacementStatus } from './upcomingReplacementStatus';

export interface UpcomingReplacement {
  filterId: number;
  filterName: string;
  cartridgeId: number;
  cartridgeName: string;
  stageNumber: number;
  nextReplacementDate: string;
  daysRemaining: number;
  status: UpcomingReplacementStatus;
}